Move cells components into components dir
