+++ context/app/static/js/components/cells/DatasetsSelectedByExpression/index.js
@@ -1,0 +1,3 @@
+import DatasetsSelectedByExpression from './DatasetsSelectedByExpression';
+
+export default DatasetsSelectedByExpression;
